requirements.txt
appdirs==1.4.4
asgiref==3.3.4
astroid==2.5.1
beautifulsoup4==4.9.3
black==20.8b1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.4.6
defusedxml==0.7.1
Django==3.2.3
django-bootstrap-v5==1.0.0
django-ckeditor==6.0.0
django-js-asset==1.2.2
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
idna==2.10
importlib-metadata==2.1.1
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.1.0
pathspec==0.8.1
Pillow==8.1.2
pycparser==2.20
PyJWT==1.7.1
pylint==2.7.2
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.2
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
wrapt==1.12.1
zipp==3.4.1
